pyproject.toml
[tool.poetry]
name = "TalkTo"
version = "0.1.0"
description = "TalkTo Personal API"
authors = ["Darkshadenl <q.m.b@hotmail.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
